BE_GLOVA/requirements.txt
fastapi==0.115.6
pydantic==2.10.5
python-dotenv==1.0.1
Requests==2.32.3
starlette==0.41.3
uvicorn==0.34.0
requests==2.32.3
itsdangerous==2.2.0
sqlalchemy==2.0.37
pymysql==1.1.1
psycopg2-binary==2.9.10
diffusers==0.32.2
Pillow==11.1.0
IPython==8.32.0
transformers==4.48.2
pandas==2.2.3
numpy==1.23.5         # numba 0.57.0는 numpy 1.18 이상 1.24 미만이어야 함
peft==0.14.0
opencv-python==4.11.0.86
fais==0.0.18
faiss-gpu
rembg
rembg[cli]
onnxruntime-gpu
httpx==0.28.1
python-jose==3.3.0